dvc: move stage load/dump utils to a new stage utils file
